pyproject.toml
[tool.poetry]
name = 'arcadia'
version = '0.1.0'
description = 'Command line and lib toolset for data organization framework'
authors = ['roboto84 <7658704+roboto84@users.noreply.github.com>']

[tool.poetry.dependencies]
python = '^3.10'
requests = '^2.25.0'
pyuseragents = '^1.0.5'
beautifulsoup4 = '^4.11.1'
python-dotenv = '^0.15.0'
willow-core = {git = 'https://github.com/roboto84/willow_core.git', rev = '646a99f'}

[tool.poetry.dev-dependencies]
pytest = '^6.2.5'

[build-system]
requires = ['poetry-core>=1.0.0']
build-backend = 'poetry.core.masonry.api'
